refactor: blog export
