+++ src/cmd/blog-export/create.ts
@@ -1,5 +1,5 @@
 import { Alert } from '@/infra/alert'
-import { BlogExportApi } from '@/service/blog-export.api'
+import { BlogExportApi } from '@/service/blog-export'
 import { BlogExportProvider } from '@/tree-view/provider/blog-export-provider'
 import { MessageItem } from 'vscode'
