+++ src/tree-view/model/blog-export/record.ts
@@ -11,7 +11,7 @@ import { BlogExportTreeItem, DownloadedExportTreeItem } from '@/tree-view/model/
 import os from 'os'
 import { escapeRegExp } from 'lodash-es'
 import { BlogExportProvider } from '@/tree-view/provider/blog-export-provider'
-import { BlogExportApi } from '@/service/blog-export.api'
+import { BlogExportApi } from '@/service/blog-export'
 
 export class BlogExportRecordTreeItem extends BaseTreeItemSource implements BaseEntryTreeItem<BlogExportTreeItem> {
     static readonly contextValue = 'cnblogs-export-record'
